[SPARK-43165][SQL] Move canWrite to DataTypeUtils
### What changes were proposed in this pull request?

Move canWrite to DataTypeUtils.

### Why are the changes needed?

canWrite access SQLConf so we can move it out from DataType to make DataType as public simpiler API.

### Does this PR introduce _any_ user-facing change?

NO

### How was this patch tested?

UT

Closes #40825 from amaliujia/catalyst_datatype_refactor_7.

Authored-by: Rui Wang <rui.wang@databricks.com>
